Split responders into separate class files
