+++ Boyfriend.csproj
@@ -21,6 +21,7 @@
     <ItemGroup>
         <PackageReference Include="DiffPlex" Version="1.7.1"/>
         <PackageReference Include="Humanizer.Core.ru" Version="2.14.1"/>
+        <PackageReference Include="JetBrains.Annotations" Version="2023.2.0"/>
         <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1"/>
         <PackageReference Include="Remora.Discord" Version="2023.3.0"/>
     </ItemGroup>
