+++ src/Commands/AboutCommandGroup.cs
@@ -1,7 +1,6 @@
 ﻿using System.ComponentModel;
 using System.Text;
 using Boyfriend.Data;
-using Boyfriend.locale;
 using Boyfriend.Services;
 using Remora.Commands.Attributes;
 using Remora.Commands.Groups;
@@ -11,9 +10,6 @@ using Remora.Discord.Commands.Feedback.Services;
 using Remora.Discord.Extensions.Embeds;
 using Remora.Discord.Extensions.Formatting;
 using Remora.Results;
-
-// ReSharper disable ClassNeverInstantiated.Global
-// ReSharper disable UnusedMember.Global
 
 namespace Boyfriend.Commands;
