+++ src/Commands/BanCommandGroup.cs
@@ -1,7 +1,6 @@
 using System.ComponentModel;
 using System.Text;
 using Boyfriend.Data;
-using Boyfriend.locale;
 using Boyfriend.Services;
 using Remora.Commands.Attributes;
 using Remora.Commands.Groups;
@@ -14,9 +13,6 @@ using Remora.Discord.Commands.Feedback.Services;
 using Remora.Discord.Extensions.Embeds;
 using Remora.Discord.Extensions.Formatting;
 using Remora.Results;
-
-// ReSharper disable ClassNeverInstantiated.Global
-// ReSharper disable UnusedMember.Global
 
 namespace Boyfriend.Commands;
