+++ src/Data/Options/IOption.cs
@@ -5,6 +5,6 @@ namespace Boyfriend.Data.Options;
 
 public interface IOption {
     string Name { get; }
-    object GetAsObject(JsonNode settings);
-    Result Set(JsonNode         settings, string from);
+    string Display(JsonNode settings);
+    Result Set(JsonNode     settings, string from);
 }
