+++ src/Data/Options/SnowflakeOption.cs
@@ -1,5 +1,5 @@
 using System.Text.Json.Nodes;
-using Boyfriend.locale;
+using Remora.Discord.Extensions.Formatting;
 using Remora.Rest.Core;
 using Remora.Results;
 
@@ -7,6 +7,10 @@ namespace Boyfriend.Data.Options;
 
 public class SnowflakeOption : Option<Snowflake> {
     public SnowflakeOption(string name) : base(name, 0UL.ToSnowflake()) { }
+
+    public override string Display(JsonNode settings) {
+        return Name.EndsWith("Channel") ? Mention.Channel(Get(settings)) : Mention.Role(Get(settings));
+    }
 
     public override Snowflake Get(JsonNode settings) {
         var property = settings[Name];
